[Fix] Fix Banner Image And PostList Width Fix
